requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
_pytorch_select=0.1=cpu_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
boost=1.72.0=py37h9de70de_0
boost-cpp=1.72.0=h9359b55_3
bzip2=1.0.8=h516909a_3
ca-certificates=2020.12.5=ha878542_0
caffe=1.0=py37h49046e7_210
cairo=1.16.0=h3fc0475_1005
certifi=2020.12.5=py37h89c1867_1
cffi=1.14.1=py37h2b28604_0
cloudpickle=1.6.0=py_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py_2
cytoolz=0.11.0=py37h8f50634_0
dask-core=2.28.0=py_0
dbus=1.13.6=he372182_0
decorator=4.4.2=py_0
expat=2.2.9=he1b5a44_2
ffmpeg=4.3.1=h167e202_0
fontconfig=2.13.1=h1056068_1002
freetype=2.10.2=he06d7ca_0
gettext=0.19.8.1=hc5be6a0_1002
gflags=2.2.2=he1b5a44_1004
glib=2.66.0=h0dae87d_0
glog=0.4.0=h49b9bf7_3
gmp=6.2.0=he1b5a44_2
gnutls=3.6.13=h79a8f9a_0
graphite2=1.3.13=he1b5a44_1001
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
h5py=2.10.0=nompi_py37h90cd8ad_104
harfbuzz=2.7.2=hee91db6_0
hdf5=1.10.6=nompi_h3c11f04_101
icu=67.1=he1b5a44_0
imageio=2.9.0=py_0
ipython=7.18.1=py37hc6149b9_0
ipython_genutils=0.2.0=py_1
jasper=1.900.1=h07fcdf6_1006
jedi=0.15.2=py37_0
joblib=0.16.0=py_0
jpeg=9d=h516909a_0
kiwisolver=1.2.0=py37h99015e2_0
krb5=1.17.1=hfafb76e_3
lame=3.100=h14c3975_1001
lcms2=2.11=hbd6801e_0
ld_impl_linux-64=2.34=hc38a660_9
leveldb=1.18=h6416369_1001
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libclang=10.0.1=default_hde54327_1
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=hcdb4288_2
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.3.0=h24d8f2e_16
libgfortran-ng=7.5.0=hdf63c60_16
libiconv=1.16=h516909a_0
liblapack=3.8.0=14_mkl
liblapacke=3.8.0=14_mkl
libllvm10=10.0.1=he513fc3_3
libopencv=4.4.0=py37_2
libpng=1.6.37=hed695b0_2
libpq=12.3=h5513abc_0
libprotobuf=3.12.4=h8b12597_0
libstdcxx-ng=9.3.0=hdf63c60_16
libtiff=4.1.0=hc7e4089_6
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=h68273f3_2
llvm-openmp=10.0.1=hc9558a2_0
lmdb=0.9.24=h516909a_0
lz4-c=1.9.2=he1b5a44_3
matplotlib=3.3.1=1
matplotlib-base=3.3.1=py37hd478181_1
mkl=2019.5=281
mkl-service=2.3.0=py37h516909a_0
mkl_fft=1.1.0=py37hc1659b7_1
mkl_random=1.0.4=py37hd81dba3_0
mysql-common=8.0.21=2
mysql-libs=8.0.21=hf3661c5_2
ncurses=6.2=he1b5a44_1
nettle=3.4.1=h1bed415_1002
networkx=2.5=py_0
ninja=1.10.1=h99015e2_0
nspr=4.29=he1b5a44_0
nss=3.57=he751ad9_0
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
olefile=0.46=py_0
opencv=4.4.0=py37_2
openh264=2.1.1=h8b12597_0
openssl=1.1.1i=h7f98852_0
pandas=1.1.1=py37h3340039_0
parso=0.5.2=py_0
patsy=0.5.1=py_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=py37hc8dfbb8_1
pickleshare=0.7.5=py37hc8dfbb8_1001
pillow=7.2.0=py37h718be6c_1
pip=20.2.2=py_0
pixman=0.38.0=h516909a_1003
prompt-toolkit=3.0.7=py_0
protobuf=3.12.4=py37h3340039_0
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py_1001
py-opencv=4.4.0=py37h43977f1_2
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.1=py_0
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.7.8=h6f2ec95_1_cpython
python-dateutil=2.8.1=py_0
python-gflags=3.1.2=py_0
python-leveldb=0.194=py37he1b5a44_1000
python_abi=3.7=1_cp37m
pytorch=1.6.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2020.1=pyh9f0ad1d_0
pywavelets=1.1.1=py37h161383b_2
pyyaml=5.3.1=py37h8f50634_0
qt=5.12.9=h1f2b2cb_0
readline=8.0=he28a2e2_2
scikit-image=0.15.0=py37hb3f55d8_2
scikit-learn=0.24.0=py37h69acf81_0
scipy=1.5.2=py37hb14ef9d_0
seaborn=0.10.1=1
seaborn-base=0.10.1=py_1
setuptools=49.6.0=py37hc8dfbb8_0
six=1.15.0=pyh9f0ad1d_0
snappy=1.1.8=he1b5a44_3
sqlite=3.33.0=h4cf870e_0
statsmodels=0.12.0=py37h8f50634_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hed695b0_0
toolz=0.11.1=py_0
torchvision=0.7.0=py37_cu101
tornado=6.0.4=py37h8f50634_1
traitlets=5.0.4=py_1
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.35.1=pyh9f0ad1d_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.12=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
zlib=1.2.11=h516909a_1007
zstd=1.4.5=h6597ccf_2
